修复：C2664

+++ Unreal/CarlaUE4/Plugins/UnrealRoboticsLab/Source/URLab/Private/MuJoCo/Utils/MjXmlUtils.cpp
@@ -72,7 +72,7 @@ bool MjXmlUtils::ReadAttrFloat(const FXmlNode* Node, const TCHAR* Attr, float& O
     return true;
 }
 
-bool MjXmlUtils::ReadAttrDouble(const FXmlNode* Node, const TCHAR* Attr, double& Out, bool& bOverride)
+bool MjXmlUtils::ReadAttrDouble(const FXmlNode* Node, const TCHAR* Attr, float& Out, bool& bOverride)
 {
     FString Str = Node->GetAttribute(Attr);
     if (Str.IsEmpty()) return false;
